+++ hahomematic/central/__init__.py
@@ -36,6 +36,7 @@ from hahomematic.const import (
     DEFAULT_ENABLE_DEVICE_FIRMWARE_CHECK,
     DEFAULT_ENABLE_PROGRAM_SCAN,
     DEFAULT_ENABLE_SYSVAR_SCAN,
+    DEFAULT_IGNORE_CUSTOM_DEVICE_DEFINITION_MODELS,
     DEFAULT_MAX_READ_WORKERS,
     DEFAULT_PERIODIC_REFRESH_INTERVAL,
     DEFAULT_PROGRAM_MARKERS,
@@ -400,7 +401,6 @@ class CentralUnit(PayloadMixin):
                 f"START: Failed to start central unit {self.name}: {reduce_args(args=oserr.args)}"
             ) from oserr
 
-        await self._parameter_visibility.load()
         if self._config.start_direct:
             if await self._create_clients():
                 for client in self._clients.values():
@@ -1665,6 +1665,7 @@ class CentralConfig:
         enable_device_firmware_check: bool = DEFAULT_ENABLE_DEVICE_FIRMWARE_CHECK,
         enable_program_scan: bool = DEFAULT_ENABLE_PROGRAM_SCAN,
         enable_sysvar_scan: bool = DEFAULT_ENABLE_SYSVAR_SCAN,
+        ignore_custom_device_definition_models: tuple[str, ...] = DEFAULT_IGNORE_CUSTOM_DEVICE_DEFINITION_MODELS,
         interfaces_requiring_periodic_refresh: tuple[Interface, ...] = INTERFACES_REQUIRING_PERIODIC_REFRESH,
         json_port: int | None = None,
         listen_ip_addr: str | None = None,
@@ -1690,7 +1691,8 @@ class CentralConfig:
         self.enable_program_scan: Final = enable_program_scan
         self.enable_sysvar_scan: Final = enable_sysvar_scan
         self.host: Final = host
-        self.interfaces_requiring_periodic_refresh = interfaces_requiring_periodic_refresh
+        self.ignore_custom_device_definition_models: Final = ignore_custom_device_definition_models
+        self.interfaces_requiring_periodic_refresh: Final = interfaces_requiring_periodic_refresh
         self.json_port: Final = json_port
         self.listen_ip_addr: Final = listen_ip_addr
         self.listen_port: Final = listen_port