+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.12.11 (2024-12-22)
 
+- Remove unignore file import
 - Rename has_markers to enabled_default
 - Use NamedTuple for datapoint key